# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "CatalogCoverageResponse": {
        "properties": {
          "configured_subjects": {
            "items": {
              "type": "string"
            },
            "title": "Configured Subjects",
            "type": "array"
          },
          "elective_subjects": {
            "items": {
              "type": "string"
            },
            "title": "Elective Subjects",
            "type": "array"
          },
          "subjects": {
            "items": {
              "$ref": "#/components/schemas/SubjectCoverage"
            },
            "title": "Subjects",
            "type": "array"
          },
          "term": {
            "title": "Term",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "term",
          "configured_subjects",
          "elective_subjects",
          "subjects",
          "warnings"
        ],
        "title": "CatalogCoverageResponse",
        "type": "object"
      },
...
      "CourseDetailResponse": {
        "properties": {
          "catalog_note": {
            "default": "Catalog coverage has not been checked. Regenerate the plan to check it.",
            "title": "Catalog Note",
            "type": "string"
          },
          "catalog_status": {
            "default": "unknown",
            "enum": [
              "present",
              "subject_not_configured",
              "course_missing",
              "unresolved",
              "unknown"
            ],
            "title": "Catalog Status",
            "type": "string"
          },
          "course_code": {
            "title": "Course Code",
            "type": "string"
          },
          "credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits"
          },
          "credits_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits Max"
          },
          "credits_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits Min"
          },
          "credits_options": {
            "items": {
              "type": "number"
            },
            "title": "Credits Options",
            "type": "array"
          },
          "credits_status": {
            "default": "unverified",
            "enum": [
              "fixed",
              "variable",
              "unverified",
              "missing"
            ],
            "title": "Credits Status",
            "type": "string"
          },
          "metadata_warnings": {
            "items": {
              "type": "string"
            },
            "title": "Metadata Warnings",
            "type": "array"
          },
          "prerequisites": {
            "items": {
              "type": "string"
            },
            "title": "Prerequisites",
            "type": "array"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionResponse"
            },
            "title": "Sections",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "title_status": {
            "default": "unverified",
            "enum": [
              "verified",
              "unverified",
              "missing"
            ],
            "title": "Title Status",
            "type": "string"
          }
        },
        "required": [
          "course_code",
          "title",
          "credits",
          "title_status",
          "credits_status",
          "credits_min",
          "credits_max",
          "credits_options",
          "metadata_warnings",
          "catalog_status",
          "catalog_note",
          "prerequisites",
          "sections"
        ],
        "title": "CourseDetailResponse",
        "type": "object"
      },
      "CourseResponse": {
        "properties": {
          "catalog_note": {
            "default": "Catalog coverage has not been checked. Regenerate the plan to check it.",
            "title": "Catalog Note",
            "type": "string"
          },
          "catalog_status": {
            "default": "unknown",
            "enum": [
              "present",
              "subject_not_configured",
              "course_missing",
              "unresolved",
              "unknown"
            ],
            "title": "Catalog Status",
            "type": "string"
          },
          "course_code": {
            "title": "Course Code",
            "type": "string"
          },
          "credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits"
          },
          "credits_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits Max"
          },
          "credits_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits Min"
          },
          "credits_options": {
            "items": {
              "type": "number"
            },
            "title": "Credits Options",
            "type": "array"
          },
          "credits_status": {
            "default": "unverified",
            "enum": [
              "fixed",
              "variable",
              "unverified",
              "missing"
            ],
            "title": "Credits Status",
            "type": "string"
          },
          "metadata_warnings": {
            "items": {
              "type": "string"
            },
            "title": "Metadata Warnings",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "title_status": {
            "default": "unverified",
            "enum": [
              "verified",
              "unverified",
              "missing"
            ],
            "title": "Title Status",
            "type": "string"
          }
        },
        "required": [
          "course_code",
          "title",
          "credits",
          "title_status",
          "credits_status",
          "credits_min",
          "credits_max",
          "credits_options",
          "metadata_warnings",
          "catalog_status",
          "catalog_note"
        ],
        "title": "CourseResponse",
        "type": "object"
      },
...
      "GenerateRequest": {
        "properties": {
          "parsed_degree": {
            "additionalProperties": true,
            "title": "Parsed Degree",
            "type": "object"
          },
          "preferences": {
            "additionalProperties": true,
            "title": "Preferences",
            "type": "object"
          }
        },
        "required": [
          "parsed_degree",
          "preferences"
        ],
        "title": "GenerateRequest",
        "type": "object"
      },
      "GeneratedPlan": {
        "properties": {
          "projected_graduation": {
            "title": "Projected Graduation",
            "type": "string"
          },
          "semesters": {
            "items": {
              "$ref": "#/components/schemas/SemesterCard"
            },
            "title": "Semesters",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "semesters",
          "projected_graduation",
          "warnings"
        ],
        "title": "GeneratedPlan",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ParseRequest": {
        "properties": {
          "client_pdf_hash": {
            "title": "Client Pdf Hash",
            "type": "string"
          },
          "pdf_base64": {
            "title": "Pdf Base64",
            "type": "string"
          }
        },
        "required": [
          "pdf_base64",
          "client_pdf_hash"
        ],
        "title": "ParseRequest",
        "type": "object"
      },
      "ParseResponse": {
        "properties": {
          "parsed": {
            "$ref": "#/components/schemas/ParsedDegreeValidated"
          },
          "server_hash": {
            "title": "Server Hash",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "parsed",
          "server_hash",
          "warnings"
        ],
        "title": "ParseResponse",
        "type": "object"
      },
...
      "SectionResponse": {
        "properties": {
          "course_code": {
            "title": "Course Code",
            "type": "string"
          },
          "crn": {
            "title": "Crn",
            "type": "string"
          },
          "meetings": {
            "items": {
              "$ref": "#/components/schemas/MeetingResponse"
            },
            "title": "Meetings",
            "type": "array"
          },
          "open_seats": {
            "title": "Open Seats",
            "type": "integer"
          },
          "professor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Professor Name"
          },
          "scraped_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scraped At"
          },
          "total_seats": {
            "title": "Total Seats",
            "type": "integer"
          }
        },
        "required": [
          "crn",
          "course_code",
          "professor_name",
          "total_seats",
          "open_seats",
          "scraped_at",
          "meetings"
        ],
        "title": "SectionResponse",
        "type": "object"
      },
...
      "SubjectCoverage": {
        "properties": {
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "course_count": {
            "title": "Course Count",
            "type": "integer"
          },
          "section_count": {
            "title": "Section Count",
            "type": "integer"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "subject",
          "configured",
          "course_count",
          "section_count"
        ],
        "title": "SubjectCoverage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NJIT Schedule Builder API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/catalog/coverage": {
      "get": {
        "operationId": "catalog_coverage_api_catalog_coverage_get",
        "parameters": [
          {
            "in": "query",
            "name": "term",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}(10|50|90)$",
              "title": "Term",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogCoverageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Catalog Coverage",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses": {
      "get": {
        "operationId": "search_courses_api_courses_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CourseResponse"
                  },
                  "title": "Response Search Courses Api Courses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Courses",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/{code}": {
      "get": {
        "operationId": "get_course_api_courses__code__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Course",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/courses/{code}/sections": {
      "get": {
        "operationId": "get_course_sections_api_courses__code__sections_get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "term",
            "required": true,
            "schema": {
              "pattern": "^\\d{4}(10|50|90)$",
              "title": "Term",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SectionResponse"
                  },
                  "title": "Response Get Course Sections Api Courses  Code  Sections Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Course Sections",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/plan/generate": {
      "post": {
        "description": "Stateless plan generation. Accepts a ParsedDegreeValidated (from /api/plan/parse)\nand student preferences, returns a semester-by-semester plan.\nNothing is stored server-side \u2014 the client persists the result to localStorage.",
        "operationId": "generate_degree_plan_api_plan_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Degree Plan",
        "tags": [
          "plan"
        ]
      }
    },
...
    "/api/plan/parse": {
      "post": {
        "description": "Stateless PDF parsing endpoint. Receives a base64-encoded DegreeWorks PDF,\nextracts structured degree data, and returns it as JSON. Nothing is stored\nserver-side \u2014 the client persists the result to localStorage.",
        "operationId": "parse_degree_works_api_plan_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parse Degree Works",
        "tags": [
          "plan"
        ]
      }
    },
    "/api/professors/{name}": {
      "get": {
        "operationId": "get_professor_api_professors__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessorResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Professor",
        "tags": [
          "courses"
        ]
      }
    },
...
  }
}